requirements.txt
django==2.2.4
libsass==0.19.2
watchdog==0.9.0
ansi2html==1.5.2
polib==1.1.0
requests==2.22.0